+++ KodiPlaylistEditor/KodiPlaylistEditor.csproj
@@ -151,6 +151,7 @@
     <Compile Include="ClassRepeatButton.cs">
       <SubType>Component</SubType>
     </Compile>
+    <Compile Include="ClassYTExplode.cs" />
     <Compile Include="CmdOutput.cs">
       <SubType>Form</SubType>
     </Compile>
